feat: plan-based billing system overhaul
Transition from flat credit model to tiered plan-based billing with 9 plans
(Trial, Basic, Starter, Business, Business Plus, Agency, Agency Plus,
Enterprise, Custom).

Backend:
- Add users.plan, users.custom_credit_price, jobs.plan_at_creation columns
- Change users.credits and jobs.cost_in_credits from Integer to Numeric(12,1)
- New backend/app/core/plans.py with all plan definitions and helpers
- Plan-aware credit deduction in enrich.py, jobs.py, and workers
- Paid plans get free enrichment/verification; trial charges 0.5 credits each
- Sales Nav scraping costs 1 credit/profile across all plans
- Plan-aware top-up pricing in payments.py (per-plan $/credit rates)
- Admin endpoints to assign plans and set custom credit prices
- New signup credits: 1000 (was 100)

Frontend:
- Interactive PricingSlider component with stepped range slider
- Plan-aware Get Credits page with plan banner and dynamic pricing
- Admin console Plan column with dropdown and custom price editor
- Replaced old pricing sections on marketing page and dashboard

Workers:
- index.js reads plan_at_creation for plan-aware deduction
- enrichment_worker.py uses job.plan_at_creation (not current user.plan)

Made-with: Cursor

+++ backend/app/schemas/enrich.py
@@ -25,5 +25,5 @@ class EnrichResponse(BaseModel):
     status: str  # valid | catchall | not_found
     pattern: Optional[str] = None
     mx_provider: Optional[str] = None
-    credits_used: int
-    credits_remaining: int
+    credits_used: float
+    credits_remaining: float
